requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
chromadb==0.4.18

# AI Models - OpenAI (commented, for future use when key is purchased)
openai==1.12.0
httpx==0.27.0

# AI Models - Gemini (currently active)
google-generativeai==0.3.2

PyPDF2==3.0.1
pdfplumber==0.10.3
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
